Rework time types. Use RTPS Timestamp (Time_t) and Duration consistently.
Do not use other time types unless really needed.
Others are std::time::Instant and Duration. The downside with these is that
Instant cannot be obtained from anywhere but system clock, or by arithmetic
from Duration/Instant.

Chrono library types are more flexible (can be converted to/from), but also
more complicated. The "time" library is a bit odd, as it artifially separates
Date and Time, which does not make sense for machine-to-machine use.