Fix deserialization issues in payload decoding
We cannot use an automatically derived Readable trait for
BuiltinCryptoTransformationKind, nor use similar traits relying on it
outside the builtin plugin, as this is not the form in which it is
transported over the wire. Instead we should deserialize to CryptoHeader
and then convert to BuiltinCryptoHeader.

Additionally, a length value is included for the data when serializing
CryptoContent, which has to be removed before decoding. This is a
temporary fix, as the way we handle payload protection needs changes.