Do not use topic qos for DataSampleCache
By the construct_topic_and_poll macro, the topic qos is the subscriber
qos, which has history depth 1. If we use that for DataSampleCache, it
will delete all but the newest entry in add_sample.

TODO: Cloning the qos from the simple datareader might not be correct
either. At the moment the way datareaders are constructed seems really
convoluted, especially the macro , the nested datareaders and the
different qos collections that sometimes override each other, and
probably deserves a closer look.