Adjust GAP messages so FastDDS always accepts them, while still complying with the spec.
Always set gapList.base as the exclusive endpoint of the contiguous range of sequence numbers in the GAP message. Before, we set gapList.base as the inclusive endpoint. This was not against the spec, but it meant that we could send messages where gap_start == gapList.base, in which case some DDS implementations (at least FastDDS) refused to process the GAP message. The new implementation also complies with the spec.